+++ buildkite/resource_test_suite_test.go
@@ -13,7 +13,6 @@ import (
 
 func TestAcc_testSuiteAddRemove(t *testing.T) {
 	t.Parallel()
-	t.Skip("Bug in Test Analytics API")
 	var suite getTestSuiteSuite
 
 	resource.Test(t, resource.TestCase{
@@ -52,7 +51,6 @@ func TestAcc_testSuiteAddRemove(t *testing.T) {
 
 func TestAcc_testSuiteUpdate(t *testing.T) {
 	t.Parallel()
-	t.Skip("Bug in Test Analytics API")
 	var suite getTestSuiteSuite
 
 	resource.Test(t, resource.TestCase{
@@ -117,7 +115,6 @@ func TestAcc_testSuiteUpdate(t *testing.T) {
 
 func TestAcc_testSuiteTeamOwnerResolution(t *testing.T) {
 	t.Parallel()
-	t.Skip("Bug in Test Analytics API")
 	var suite getTestSuiteSuite
 
 	resource.Test(t, resource.TestCase{